various improvements around effects
